Cleanup MCKillSwitchAlert code so it's harder to break it
If the app hosting the kill switch tried to present a view controller at the same time the kill switch alert was presented/dismissed, weird behaviour would occur in the app.

Add `MCKillSwitchAlertController` that will be used to know if the kill switch is presented and that will prevent the app from presenting something on top of it (by throwing an exception).

Also cleaned the code by removing a dead function (`destroyAlertView`) and passing the button to execute instead of finding it again by index after recreating the array.

Finally, fixed the `SKStoreProductViewController` that would not be presented if something was already being presented over the root view controller.